Moved all type traits out of Util.h and into their own header file.
