+++ mod_dev_tool/common/inc/Util.h
@@ -20,6 +20,7 @@
 # include "Types.h"
 # include "Logger.h"
 # include "PreprocessorUtils.h"
+# include "TypeTraits.h"
 
 namespace HMDT {
     // Forward declare this, as we don't need to include the whole file yet.
@@ -472,36 +473,6 @@ namespace HMDT {
     }
 
     /**
-     * @brief Implements the C++20 std::type_identity
-     */
-    template<typename T>
-    struct Identity {
-        using type = T;
-    };
-    template<typename T>
-    using Identity_t = typename Identity<T>::type;
-
-    // https://stackoverflow.com/a/39492671
-    /**
-     * @brief Removes all pointer indirections on the given type.
-     */
-    template<typename T>
-    struct RemoveAllPointers:
-        std::conditional_t<std::is_pointer_v<T>,
-                           RemoveAllPointers<std::remove_pointer_t<T>>,
-                           Identity<T>
-                          >
-    { };
-    template<typename T>
-    using RemoveAllPointers_t = typename RemoveAllPointers<T>::type;
-
-/// @cond
-    template<typename T>
-    T __void_if_no_value(T);
-    [[maybe_unused]] void __void_if_no_value();
-/// @endcond
-
-    /**
      * @brief Runs the given function when the current scope ends
      */
     struct RunAtScopeEnd {
@@ -524,12 +495,6 @@ namespace HMDT {
      */
     template<typename T>
     using Ref = std::reference_wrapper<T>;
-
-    /**
-     * @brief Helper type for static_assert which always evaluates to false.
-     */
-    template<typename...>
-    constexpr std::false_type alwaysFalse{};
 }
 
 #endif
